Small change to make code compile with Windows' math.h
git-svn-id: https://csp.svn.sourceforge.net/svnroot/csp/trunk@1858 7152ba15-5bd8-0310-b828-e896570a432b

+++ csp/cspsim/f16/StoresManagementPages.cpp
@@ -622,7 +622,7 @@ protected:
 			if (index == 1) {
 				if (!m_Current.empty()) {
 					double value = std::min(entry.high, std::max<unsigned>(entry.low, atoi(m_Current.c_str())));
-					entry.value = value / pow(10.0, entry.decimal);
+					entry.value = value / pow(10.0, (double)entry.decimal);
 					m_Current = "";
 					formatOutput(m_Index);
 				}
